% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots2pane.R
\name{plots2pane}
\alias{plots2pane}
\title{plots2pane}
\usage{
plots2pane(plt_list, rows, cols)
}
\arguments{
\item{plt_list}{List of ggplot objects.}

\item{rows}{Rows of pane.}

\item{cols}{Columns of pane.}
}
\description{
Generates an aligned, multi-plot pane from a list of ggplots objects.
}
\examples{
x = c(1,2,3)
y = c(1,2,3)
dataset = data.frame(cbind(x,y))
plots = list()
plots[[1]] = ggplot(dataset, aes(x=x, y=y)) + geom_line()
plots[[2]] = ggplot(dataset, aes(x=x, y=y)) + geom_line()
plots2pane(plots, 1, 2)

}
